crudit tehty loppuun

+++ app/views.py
@@ -1,5 +1,30 @@
 from django.shortcuts import render, redirect
 from .models import Supplier, Product
+from django.contrib.auth import authenticate, login, logout
+
+
+# LANDING AFTER LOGIN
+
+def loginview(request):
+    return render (request, "loginpage.html")
+
+def login_action(request):
+    user = request.POST['username']
+    passw = request.POST['password']
+    user = authenticate(username = user, password = passw)
+    if user:
+        login(request, user)
+        context = {'name': user}
+        return render(request,'landingpage.html',context)
+    else:
+        return render(request, 'loginerror.html')
+
+def logout_action(request):
+    logout(request)
+    return render(request, 'loginpage.html')
+
+
+# After Login    
 
 def landingview(request):
     return render (request, 'landingpage.html')
@@ -7,9 +32,21 @@ def landingview(request):
 # SUPPLIERS
 
 def supplierlistview(request):
-    supplierlist = Supplier.objects.all()
-    context = {'suppliers': supplierlist}
-    return render (request,"suppliers.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        supplierlist = Supplier.objects.all()
+        context = {'suppliers': supplierlist}
+        return render (request,"suppliers.html",context)
+
+def searchsuppliers(request):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        search = request.POST['search']
+        filtered = Supplier.objects.filter(companyname__icontains=search)
+        context = {'suppliers': filtered}
+        return render (request,"suppliers.html",context)
 
 def addsupplier(request):
     a = request.POST['companyname']
@@ -31,40 +68,52 @@ def edit_supplier_get(request, id):
     return render (request,"edit_supplier.html",context)
 
 def edit_supplier_post(request, id):
-    item = Supplier.objects.get(id = id)
-    item.contactname = request.POST['contactname']
-    item.address = request.POST['address']
-    item.phone = request.POST['phone']
-    item.email = request.POST['email']
-    item.country = request.POST['country']
-    item.save()
-    return redirect(supplierlistview)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        item = Supplier.objects.get(id = id)
+        item.contactname = request.POST['contactname']
+        item.address = request.POST['address']
+        item.phone = request.POST['phone']
+        item.email = request.POST['email']
+        item.country = request.POST['country']
+        item.save()
+        return redirect(supplierlistview)
 
 
 
 # PRODUCTS
 
 def productlistview(request):
-    productlist = Product.objects.all()
-    supplierlist = Supplier.objects.all()
-    context = {'products': productlist, 'suppliers': supplierlist}
-    return render (request,"products.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        productlist = Product.objects.all()
+        supplierlist = Supplier.objects.all()
+        context = {'products': productlist, 'suppliers': supplierlist}
+        return render (request,"products.html",context)
 
 
 def addproduct(request):
-    a = request.POST['productname']
-    b = request.POST['packagesize']
-    c = request.POST['unitprice']
-    d = request.POST['unitsinstock']
-    e = request.POST['supplier']
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        a = request.POST['productname']
+        b = request.POST['packagesize']
+        c = request.POST['unitprice']
+        d = request.POST['unitsinstock']
+        e = request.POST['supplier']
     
     Product(productname = a, packagesize = b, unitprice = c, unitsinstock = d, supplier = Supplier.objects.get(id = e)).save()
     return redirect(request.META['HTTP_REFERER'])
 
 
 def deleteproduct(request, id):
-    Product.objects.filter(id = id).delete()
-    return redirect(request.META['HTTP_REFERER'])
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        Product.objects.filter(id = id).delete()
+        return redirect(request.META['HTTP_REFERER'])
 
 def edit_product_get(request, id):
     product = Product.objects.filter(id = id)
